% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcofnpkg-package.R
\docType{package}
\name{rcofnpkg-package}
\alias{rcofnpkg}
\alias{rcofnpkg-package}
\title{rcofnpkg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DanielQuiroz97/rcofnpkg}
  \item Report bugs at \url{https://github.com/DanielQuiroz97/rcofnpkg/issues}
}

}
\author{
\strong{Maintainer}: Daniel Quiroz \email{cristianquirozd1997@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
