% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matches.R
\name{uss_make_matches}
\alias{uss_make_matches}
\title{Makes standards datagramse}
\usage{
uss_make_matches(data_engsoc)
}
\arguments{
\item{data_engsoc}{data.frame from engsoc soccer data}
}
\value{
data.frame, tibble with prcessed data
}
\description{
Makes standards datagramse
}
\examples{
uss_make_matches(engsoccerdata::italy)
}
